requirements.txt
nltk
spacy
pyresparser
pandas
python-dateutil
datetime
python-docx
PyMuPDF
openpyxl
pdfminer.six
